Use Maven
This makes dependency management much easier: instead of making sure that
you have Fiji installed and updated at /Applications/Fiji.app/ (which is
not the best location on Linux and impossible on Windows), you just need
to clone this project and run

	mvn

That's it.

Making *any* project a Maven one is easy:

* move your sources to src/main/java/
* move the resources (such as plugins.config) to src/main/resources/
* write a pom.xml

The most primitive pom.xml for an ImageJ 1.x plugin looks like this (the
first 5 lines are just boilerplate, i.e. saying "Hi Maven, it's me, the
pom"):

-- snip --
<?xml version="1.0" encoding="UTF-8"?>
<project xmlns="http://maven.apache.org/POM/4.0.0"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
                http://maven.apache.org/xsd/maven-4.0.0.xsd">
        <modelVersion>4.0.0</modelVersion>

	<groupId>my.group</groupId>
	<artifactId>My_Plugin</artifactId>
	<version>1.0.0</version>

	<dependencies>
		<dependency>
			<groupId>net.imagej</groupId>
			<artifactId>ij</artifactId>
			<version>1.44</version>
		</dependency>
	</dependencies>

	<repositories>
		<repository>
			<id>imagej.releases</id>
			<url>http://maven.imagej.net/content/repositories/releases</url>
		</repository>
	</repositories>
</project>
-- snap --

The pom.xml I included here is slightly more complicated because it uses
the scijava (http://scijava.org/) pom as parent; that way, current
versions are available as ${imagej1.version} and friends.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>